peniko

Enum ImageFormat

#[non_exhaustive]

#[repr(u8)]
pub enum ImageFormat {
    Rgba8 = 0,
    Bgra8 = 1,
}

Defines the pixel format of an image.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Rgba8 = 0

32-bit RGBA with 8-bit channels.

Bgra8 = 1

32-bit BGRA with 8-bit channels.

Implementations

impl ImageFormat

pub fn size_in_bytes(self, width: u32, height: u32) -> Option<usize>

Returns the required size in bytes for an image in this format of the given dimensions.

A result of None indicates an overflow in the size calculation.