peniko

Module gradient

Structs

ColorStop

Offset and color of a transition point in a gradient.

ColorStops

Collection of color stops.

Gradient

Definition of a gradient that transitions between two or more colors.

LinearGradientPosition

Parameters that define the position of a linear gradient.

RadialGradientPosition

Parameters that define the position of a radial gradient.

SweepGradientPosition

Parameters that define the position of a sweep gradient.

Enums

GradientKind

Properties for the supported gradient types.

InterpolationAlphaSpace

Defines how color channels should be handled when interpolating between transparent colors.

Constants

DEFAULT_GRADIENT_COLOR_SPACE 🔒

The default for Gradient::interpolation_cs.

Traits

ColorStopsSource

Trait for types that represent a source of color stops.