peniko::image

Struct ImageSampler

pub struct ImageSampler {
    pub x_extend: Extend,
    pub y_extend: Extend,
    pub quality: ImageQuality,
    pub alpha: f32,
}

Parameters which specify how to sample an image during rendering.

When a renderer is drawing an image, they will (in most cases) not directly copy the bytes from the source image to their render target; instead, they will sample from the image. This involves determining from which part of the source image to read, and how to handle cases where the source image’s pixels are not aligned with the render target exactly, in any combination of scale, position or rotation. They might also perform an alpha multiplication, as done here. This struct contains the parameters used by sampling.

Fields
x_extend: Extend

Extend mode in the horizontal direction.

y_extend: Extend

Extend mode in the vertical direction.

quality: ImageQuality

Hint for desired rendering quality.

alpha: f32

An additional alpha multiplier to use with the image.

Implementations

impl ImageSampler

pub fn new() -> Self

Creates a new ImageSampler with default values

pub fn with_extend(self, mode: Extend) -> Self

Builder method for setting the image extend mode in both directions.

pub fn with_x_extend(self, mode: Extend) -> Self

Builder method for setting the image extend mode in the horizontal direction.

pub fn with_y_extend(self, mode: Extend) -> Self

Builder method for setting the image extend mode in the vertical direction.

pub fn with_quality(self, quality: ImageQuality) -> Self

Builder method for setting a hint for the desired image quality when rendering.

pub fn with_alpha(self, alpha: f32) -> Self

Returns the image with the alpha multiplier set to alpha.

pub fn multiply_alpha(self, alpha: f32) -> Self

Returns the image with the alpha multiplier multiplied again by alpha. The behaviour of this transformation is undefined if alpha is negative.