peniko

Struct ImageBrush

pub struct ImageBrush<D = ImageData> {
    pub image: D,
    pub sampler: ImageSampler,
}

Describes the image content of a filled or stroked shape.

This type is generic over the storage used for the image data. By default, the generic parameter is ImageData, which is a shared image with dynamic lifetime. However, different renderers can use different types here, such as a pre-registered id.

Fields
image: D

The image to render.

sampler: ImageSampler

Parameters which specify how to sample from the image during rendering.

Implementations

impl<D> ImageBrush<D>

pub fn with_extend(self, mode: Extend) -> Self

Builder method for setting the image extend mode in both directions.

pub fn with_x_extend(self, mode: Extend) -> Self

Builder method for setting the image extend mode in the horizontal direction.

pub fn with_y_extend(self, mode: Extend) -> Self

Builder method for setting the image extend mode in the vertical direction.

pub fn with_quality(self, quality: ImageQuality) -> Self

Builder method for setting a hint for the desired image quality when rendering.

pub fn with_alpha(self, alpha: f32) -> Self

Returns the image with the alpha multiplier set to alpha.

pub fn multiply_alpha(self, alpha: f32) -> Self

Returns the image with the alpha multiplier multiplied again by alpha. The behaviour of this transformation is undefined if alpha is negative.

impl ImageBrush

pub fn new(image: ImageData) -> Self

Creates a new ImageBrush for the specified ImageData with default ImageSampler.

pub fn as_ref(&self) -> ImageBrushRef<'_>

Converts an owned ImageBrush into a borrowed ImageBrushRef.

impl ImageBrush<&ImageData>

pub fn to_owned(&self) -> ImageBrush

Converts the ImageBrushRef to an owned ImageBrush.